comments and catch errors

+++ ffmpeg_library_optimizer.py
@@ -79,16 +79,19 @@ def check_optimized(item):
         return True
     else:
         print(color('file is not optimized', fg='red'))
-        
+
+#print a list of file names        
 if args.list:
     print('args list')
     for item in get_files():
         print(item)
-    
+
+#print a table of file names with codec data     
 if args.data:
     print('args data')
     print tabulate(get_data(get_files()))
 
+#check if mp4 and if not check the codes, if they are compliant, copy the streams
 if args.container:
     print('arg container')
     for item in get_data(get_files()):
@@ -98,14 +101,16 @@ if args.container:
             if(check_codecs(item)):
                 #generate temp file name
                 tempfile = item['path'][:-4] + "_temp" + item['path'][-4:]
+                #generate new file name
+                newfile = item['path'][:-4] + '.mp4'
                 #rename the origional file with _temp
                 os.rename(item['path'], tempfile)
                 #change the container, copy the streams
                 try:
-                    subprocess.check_output('ffmpeg -loglevel info -y -i "' + tempfile + '" -c:v copy -c:a copy -movflags faststart "' + item['path'][:-4] + '.mp4"')
+                    subprocess.check_output('ffmpeg -loglevel info -y -i "' + tempfile + '" -c:v copy -c:a copy -movflags faststart "' + newfile + '"')
                 except: 
                     print(color('could not change container', fg='red'))
-                if os.path.isfile(item['path'][:-4] + '.mp4'):
+                if os.path.isfile(newfile):
                     try:
                         os.remove(tempfile)
                     except:
@@ -113,43 +118,70 @@ if args.container:
                 else:
                     os.rename(tempfile, item['path'])
 
+#check the moov atom of mpv files to ensire it is at the start of the file for quick streaming
 if args.optimize:
     print('args optimize')
     for item in get_data(get_files()):
+        #check if mp4
         if(check_mp4(item)):
+            #check if not alreay optimized
             if not (check_optimized(item)):
+                #generate temp file name
                 tempfile = item['path'][:-4] + "_temp" + item['path'][-4:]
+                #generate new file name
+                newfile = item['path'][:-4] + '.mp4'
+                #rename the origional file with _temp
                 os.rename(item['path'], tempfile)
-                #
+                #copy the streams and move the moov atom 
                 try:
-                    subprocess.check_output('ffmpeg -loglevel info -y -i "' + tempfile + '" -c:v copy -c:a copy -movflags faststart "' + item['path'] + '"')
+                    subprocess.check_output('ffmpeg -loglevel info -y -i "' + tempfile + '" -c:v copy -c:a copy -movflags faststart "' + newfile + '"')
                 except: 
                     print(color('could not optimize file', fg='red'))
-                if os.path.isfile(item['path']):
-                    os.remove(tempfile)
+                #if new file exists delete the temp file
+                if os.path.isfile(newfile):
+                    try:
+                        os.remove(tempfile)
+                    except:
+                        continue
                 else:
                     os.rename(tempfile, item['path'])
-                    
+
+#check if file needs transcoding, if it does, transcode, otherwise, copy the streams                    
 if args.transcode:
     print('arg transcode')
     for item in get_data(get_files()):
-        if not (check_codecs(item)):
+        #if item is already mp4 and correctly transcoded, skip to next file
+        if (check_mp4(item) and check_codecs(item)):
+            continue
+        else:
+            #generate temp file name
             tempfile = item['path'][:-4] + "_temp" + item['path'][-4:]
+            #generate new file name
+            newfile = item['path'][:-4] + '.mp4'
+            #rename the origional file with _temp
             os.rename(item['path'], tempfile)
+            #check the video coded and copy the stream if h264
             if(item['vcodec'] == 'h264'):
                 vcodec = 'copy'
             else:
                 vcodec = 'h264'
+            #check the audio stream and copy the stream if aac/mp3
             if(item['acodec'] == 'mp3' or item['acodec'] == 'aac'):
                 acodec = 'copy'
             else:
                 acodec = 'aac'
+            #transcode the file
             try:
-                subprocess.check_output('ffmpeg -loglevel info -y -i "' + tempfile + '" -c:v ' + vcodec + ' -c:a ' + acodec + ' -preset veryfast -movflags faststart -r 24 "' + item['path'][:-4] + '.mp4"')
+                subprocess.check_output('ffmpeg -loglevel info -y -i "' + tempfile + '" -c:v ' + vcodec + ' -c:a ' + acodec + ' -preset veryfast -movflags faststart -r 24 "' + newfile + '"')
             except: 
                 print(color('could not transcode file', fg='red'))
-            if os.path.isfile(item['path'][:-4] + '.mp4'):
-                os.remove(tempfile)
+            #if new file exists delete the temp file
+            if os.path.isfile(newfile):
+                try:
+                    os.remove(tempfile)
+                except:
+                    continue
+            #if new file doesn't exist put back the old file
             else:
                 os.rename(tempfile, item['path'])
